+++ cmd/fofa/cmd/dump.go
@@ -42,7 +42,7 @@ var dumpCmd = &cli.Command{
 		&cli.StringFlag{
 			Name:        "outFile",
 			Aliases:     []string{"o"},
-			Usage:       "if not set, wirte to stdout",
+			Usage:       "if not set, write to stdout",
 			Destination: &outFile,
 		},
 		&cli.StringFlag{