+++ host.go
@@ -4,6 +4,9 @@ import (
 	"context"
 	"encoding/base64"
 	"errors"
+	"fmt"
+	"github.com/Knetic/govaluate"
+	"github.com/expr-lang/expr"
 	"math"
 	"strconv"
 	"strings"
@@ -44,10 +47,14 @@ type HostStatsData struct {
 
 // SearchOptions options of search, for post processors
 type SearchOptions struct {
-	FixUrl    bool   // each host fix as url, like 1.1.1.1,80 will change to http://1.1.1.1, https://1.1.1.1:8443 will no change
-	UrlPrefix string // default is http://
-	Full      bool   // search result for over a year
-	UniqByIP  bool   // uniq by ip
+	FixUrl     bool   // each host fix as url, like 1.1.1.1,80 will change to http://1.1.1.1, https://1.1.1.1:8443 will no change
+	UrlPrefix  string // default is http://
+	Full       bool   // search result for over a year
+	UniqByIP   bool   // uniq by ip
+	IsActive   bool   // probe website is existed, add isActive field
+	DeWildcard int    // number of wildcard domains retained
+	Filter     string // filter data by rules
+	DedupHost  bool   // prioritize subdomain data retention
 }
 
 // fixHostToUrl 替换host为url
@@ -76,6 +83,31 @@ func fixHostToUrl(res [][]string, fields []string, hostIndex int, urlPrefix stri
 		newRes = append(newRes, newRow)
 	}
 	return newRes
+}
+
+func getParamIndexThenAdd(fields []string, field string) (int, []string) {
+	paramIndex := -1
+	for index, f := range fields {
+		if f == field {
+			paramIndex = index
+			break
+		}
+	}
+	if paramIndex == -1 {
+		fields = append(fields, field)
+		paramIndex = len(fields) - 1
+	}
+	return paramIndex, fields
+}
+
+func extractVariables(filter string) ([]string, error) {
+	f, err := govaluate.NewEvaluableExpression(filter)
+	if err != nil {
+		return nil, err
+	}
+
+	variables := f.Vars()
+	return variables, nil
 }
 
 // fixUrlCheck 检查参数，构建新的field和记录相关字段的偏移
@@ -149,11 +181,21 @@ func (c *Client) postProcess(res [][]string, fields []string,
 // fields of fofa host search
 // options for search
 func (c *Client) HostSearch(query string, size int, fields []string, options ...SearchOptions) (res [][]string, err error) {
-	var full bool
-	var uniqByIP bool
+	var (
+		full        bool
+		uniqByIP    bool
+		isActive    bool
+		dedupCname  int
+		isSubDomain bool
+		filter      string
+	)
 	if len(options) > 0 {
 		full = options[0].Full
 		uniqByIP = options[0].UniqByIP
+		isActive = options[0].IsActive
+		dedupCname = options[0].DeWildcard
+		filter = options[0].Filter
+		isSubDomain = options[0].DedupHost
 	}
 
 	freeSize := c.freeSize()
@@ -198,18 +240,51 @@ func (c *Client) HostSearch(query string, size int, fields []string, options ...
 
 	uniqIPMap := make(map[string]bool)
 	// 确认fields包含ip
-	ipIndex := -1
+	var ipIndex = -1
 	if uniqByIP {
-		for index, f := range fields {
-			if f == "ip" {
-				ipIndex = index
-				break
-			}
+		ipIndex, fields = getParamIndexThenAdd(fields, "ip")
+	}
+
+	var activeSlice []string
+	// 确认fields包含link
+	var linkIndex = -1
+	if isActive {
+		linkIndex, fields = getParamIndexThenAdd(fields, "link")
+	}
+
+	dedupCnameMap := make(map[string]int)
+	// 确认fields包含ip、port、domain、title、fid
+	var portIndex, domainIndex, titleIndex, fidIndex int = -1, -1, -1, -1
+	if dedupCname > 0 {
+		ipIndex, fields = getParamIndexThenAdd(fields, "ip")
+		portIndex, fields = getParamIndexThenAdd(fields, "port")
+		domainIndex, fields = getParamIndexThenAdd(fields, "domain")
+		titleIndex, fields = getParamIndexThenAdd(fields, "title")
+		fidIndex, fields = getParamIndexThenAdd(fields, "fid")
+	}
+
+	// 过滤器配置
+	filterIndexs := make(map[string]int)
+	if len(filter) > 0 {
+		var variables []string
+		variables, err = extractVariables(filter)
+		if err != nil {
+			return nil, err
 		}
-		if ipIndex == -1 {
-			fields = append(fields, "ip")
-			ipIndex = len(fields) - 1
+
+		var filterIndex = -1
+		for _, filterField := range variables {
+			filterIndex, fields = getParamIndexThenAdd(fields, filterField)
+			filterIndexs[filterField] = filterIndex
 		}
+	}
+
+	isSubDomainMap := make(map[string][]string)
+	// 确认fields包含type
+	typeIndex := -1
+	if isSubDomain {
+		typeIndex, fields = getParamIndexThenAdd(fields, "type")
+		linkIndex, fields = getParamIndexThenAdd(fields, "link")
 	}
 
 	// 分页取数据
@@ -262,16 +337,53 @@ func (c *Client) HostSearch(query string, size int, fields []string, options ...
 						}
 						uniqIPMap[newSlice[ipIndex]] = true
 					}
+					if dedupCname > 0 {
+						key := fmt.Sprintf("%s:%s:%s:%s:%s", newSlice[ipIndex], newSlice[portIndex],
+							newSlice[domainIndex], newSlice[titleIndex], newSlice[fidIndex])
+						if _, ok := dedupCnameMap[key]; ok && dedupCnameMap[key] > 3 {
+							continue
+						}
+						dedupCnameMap[key]++
+					}
+					if len(filter) > 0 {
+						env := make(map[string]interface{})
+						for field, index := range filterIndexs {
+							env[field] = newSlice[index]
+						}
+
+						program, err := expr.Compile(filter, expr.Env(env))
+						if err != nil {
+							return nil, err
+						}
+
+						match, err := expr.Run(program, env)
+						if err != nil {
+							return nil, err
+						}
+
+						if !match.(bool) {
+							continue
+						}
+					}
+					if isActive {
+						active := CheckActive(newSlice[linkIndex])
+						activeSlice = append(activeSlice, fmt.Sprintf("%t", active))
+					}
 					results = append(results, newSlice)
 				} else if vStr, ok := result.(string); ok {
 					// 确定第一个就是ip
+					newSlice := []string{vStr}
 					if uniqByIP && ipIndex == 0 {
 						if _, ok := uniqIPMap[vStr]; ok {
 							continue
 						}
 						uniqIPMap[vStr] = true
 					}
-					results = append(results, []string{vStr})
+					if isActive && linkIndex == 0 {
+						active := CheckActive(vStr)
+						activeSlice = append(activeSlice, fmt.Sprintf("%t", active))
+					}
+					results = append(results, newSlice)
 				}
 			}
 		} else {
@@ -297,8 +409,36 @@ func (c *Client) HostSearch(query string, size int, fields []string, options ...
 		page++ // 翻页
 	}
 
+	if isSubDomain {
+		var result [][]string
+		for _, row := range res {
+			exist, found := isSubDomainMap[row[linkIndex]]
+			if found {
+				if row[linkIndex] == "" {
+					result = append(result, row)
+					continue
+				}
+				if !(exist[typeIndex] == "service" && row[typeIndex] == "subdomain") {
+					continue
+				}
+			}
+			isSubDomainMap[row[linkIndex]] = row
+		}
+
+		for _, v := range isSubDomainMap {
+			result = append(result, v)
+		}
+
+		res = result
+	}
+
 	// 后处理
 	res = c.postProcess(res, fields, hostIndex, protocolIndex, rawFieldSize, options...)
+	if isActive {
+		for index := range res {
+			res[index] = append(res[index], activeSlice[index])
+		}
+	}
 
 	return
 }